requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
async-property==0.2.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
deprecation==2.1.0
Django==5.1.3
django-cors-headers==4.6.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.2
pycparser==2.22
PyJWT==2.9.0
python-keycloak==4.7.1
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
